Implements the ability to image different polarization products.
The imager constructor now accepts an additional argument of type `Polarization`, defined in `gridding.hpp`.
`Polarization` is an `enum class` which can takes the following self-explanatory values: `XX`, `YY`, and `I`.

The value is saved in a class member attribute, and it is then passed down to the gridding routines.

In the gridding kernels, we check which polarization product we want to grid. `XX` is a complex value
whose components are at offset 0 and 1 in a baseline correlation products matrix. `YY` complex
components are at position 6 and 7. For Stokes I, the code adds up the correspoding real and imaginary
parts before gridding.

A note on the `gridding.hpp` file. I have introduced it because in the future I intend to decouple
the CPU gridding routines from the imager class, pretty much like I did for the GPU versions.
See `src/gridding_gpu.*`.