Reformat imager code to enable long exposure imaging imode.
The imager has now three public methods, `run`, `grid`, and `image`.
The `run` method simply calls `grid` and then `image`, implementing
the default behaviour the imager had up until this commit. That is,
visibilities are gridded and then imaged straight away.

Now, the grid method can be called multiple times, allowing the
caller to grid multiple coarse channels and seconds on the same
grid before it is imaged. This implements the long exposure imaging
feature.

There is still work to do. Different intervals and fine channels
in the same .dat file, or xcorr structure, are gridded on separate
grids. In long exposure mode, we only want to grid on a single grid.
Hence, we need a constructor arugment called `long_exposure` that
signals the imager class that it should run in long exposure mode.

Other changes to the code comprise moving arguments of `run_imager`
to the constructor, such as `n_pixels`.